Switch button is now an image! + few features
Android app now has a login activity, which leads to
a activity that has a ImageButton for switching the
relay on/off. A TextView has been added which will
later show server date and time (currently working on it).

On the ESP code, function for checking relay state has been added.
The function and URL for changing the relay state has been renamed.