Added dialog for configuring each device. The dialog pops up by long pressing a device icon.
The dialog is actually an Activity.

For now the start time shows the current time and the end time shows
current time + 1h. Also clicking on the date which displays "Today" by
default will open up the MaterialDatePicker.

Added dimens.xml for specifing dimensions, currently not used.

The theme looks horrible which will be changed.